# agent/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ClickHouse Operations Agent",
        "title": "ClickHouse Operations API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate user and return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get information about the currently authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Register a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/processes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns currently running queries from system.processes for a specific node",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "Get current processes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ClickHouse node hostname",
                        "name": "node",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProcessListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/processes/kill": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Kills a running query by query_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "Kill a process",
                "parameters": [
                    {
                        "description": "Kill process request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.KillProcessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KillProcessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/processes/stream": {
            "get": {
                "description": "Streams current running queries from system.processes for a specific node",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "Stream processes via SSE",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node name",
                        "name": "node",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token (alternative to Authorization header for SSE)",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/query-log": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns ClickHouse system.query_log entries filtered by time range, user, and node",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query-log"
                ],
                "summary": "List query log entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Relative window (10s|30s|1m|5m|15m)",
                        "name": "last",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start timestamp (RFC3339 or 2006-01-02 15:04:05)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End timestamp (RFC3339 or 2006-01-02 15:04:05)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Initial or effective ClickHouse user",
                        "name": "user",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ClickHouse node hostname",
                        "name": "node",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (max 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QueryLogResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid credentials"
                },
                "message": {
                    "type": "string",
                    "example": "The provided username or password is incorrect"
                }
            }
        },
        "models.KillProcessRequest": {
            "type": "object",
            "required": [
                "query_id"
            ],
            "properties": {
                "node": {
                    "type": "string",
                    "example": "primary"
                },
                "query_id": {
                    "type": "string",
                    "example": "abc-123-def"
                }
            }
        },
        "models.KillProcessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Query killed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "models.Process": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "client_name": {
                    "type": "string"
                },
                "client_version": {
                    "type": "string"
                },
                "current_database": {
                    "type": "string"
                },
                "elapsed": {
                    "description": "seconds",
                    "type": "number"
                },
                "memory_usage": {
                    "type": "integer"
                },
                "node": {
                    "type": "string"
                },
                "os_user": {
                    "type": "string"
                },
                "profile_events": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "query": {
                    "type": "string"
                },
                "query_duration_ms": {
                    "type": "integer"
                },
                "query_id": {
                    "type": "string"
                },
                "query_start_time": {
                    "description": "RFC3339",
                    "type": "string"
                },
                "read_bytes": {
                    "type": "integer"
                },
                "read_rows": {
                    "type": "integer"
                },
                "settings": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "thread_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "total_rows_approx": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                },
                "written_bytes": {
                    "type": "integer"
                },
                "written_rows": {
                    "type": "integer"
                }
            }
        },
        "models.ProcessListResponse": {
            "type": "object",
            "properties": {
                "node": {
                    "type": "string"
                },
                "processes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Process"
                    }
                }
            }
        },
        "models.QueryLogEntry": {
            "type": "object",
            "properties": {
                "client_hostname": {
                    "type": "string"
                },
                "databases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "duration_ms": {
                    "type": "integer"
                },
                "event_time": {
                    "type": "string"
                },
                "event_time_microseconds": {
                    "type": "string"
                },
                "exception": {
                    "type": "string"
                },
                "exception_code": {
                    "type": "integer"
                },
                "initial_user": {
                    "type": "string"
                },
                "memory_usage": {
                    "type": "integer"
                },
                "node": {
                    "type": "string"
                },
                "query_id": {
                    "type": "string"
                },
                "query_kind": {
                    "type": "string"
                },
                "query_text": {
                    "type": "string"
                },
                "read_bytes": {
                    "type": "integer"
                },
                "read_rows": {
                    "type": "integer"
                },
                "result_bytes": {
                    "type": "integer"
                },
                "result_rows": {
                    "type": "integer"
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "type": "string"
                },
                "written_bytes": {
                    "type": "integer"
                },
                "written_rows": {
                    "type": "integer"
                }
            }
        },
        "models.QueryLogPagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "range": {
                    "$ref": "#/definitions/models.QueryLogRange"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.QueryLogRange": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "preset": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "models.QueryLogResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.QueryLogEntry"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/models.QueryLogPagination"
                }
            }
        },
        "models.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "securepass123"
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3,
                    "example": "newuser"
                }
            }
        },
...
        "models.TokenResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer",
                    "example": 3600
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "models.UserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}