# agent/internal/api/v1/router.go
package v1

import (
	"time"

	"clickhouse-ops/internal/api/auth"
	_ "clickhouse-ops/internal/api/docs" // swagger docs - generated by swag init
	"clickhouse-ops/internal/api/middleware"
	"clickhouse-ops/internal/api/v1/handlers"
	"clickhouse-ops/internal/config"
	"clickhouse-ops/internal/logger"

	"github.com/gin-gonic/gin"
)

// RouterConfig holds router configuration
type RouterConfig struct {
	JWTSecretKey     string
	JWTTokenDuration time.Duration // token duration
	RateLimitRPS     float64
	RateLimitBurst   int
	Logger           *logger.Logger
	Config           *config.Config
}

// SetupRouter sets up API v1 routes
func SetupRouter(cfg RouterConfig) *gin.Engine {
	router := gin.New()

	// Global middleware
	router.Use(gin.Recovery())                           // Panic recovery
	router.Use(middleware.TracingMiddleware())           // Tracing
	router.Use(middleware.LoggingMiddleware(cfg.Logger)) // Logging

	// Optional rate limiting
	if cfg.RateLimitRPS > 0 {
		rateLimiter := middleware.NewRateLimiter(cfg.RateLimitRPS, cfg.RateLimitBurst)
		router.Use(rateLimiter.RateLimitMiddleware())
	}

	// CORS middleware
	router.Use(func(c *gin.Context) {
		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Content-Type, Authorization, X-Trace-ID")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	})

	// Initialize JWT manager
	jwtManager := auth.NewJWTManager(
		cfg.JWTSecretKey,
		cfg.JWTTokenDuration,
	)

	// Initialize handlers
	authHandler, err := handlers.NewAuthHandler(jwtManager, cfg.Logger)
	if err != nil {
		cfg.Logger.Errorf("Failed to initialize auth handler: %v", err)
	}
	healthHandler := handlers.NewHealthHandler()
	metricsHandler, err := handlers.NewMetricsHandler(cfg.Logger, cfg.Config)
	if err != nil {
		cfg.Logger.Errorf("Failed to initialize metrics handler: %v", err)
		// Continue without metrics handler - endpoints will return errors
	}
	queryLogHandler, err := handlers.NewQueryLogHandler(cfg.Logger, cfg.Config)
	if err != nil {
		cfg.Logger.Errorf("Failed to initialize query log handler: %v", err)
	}
	processHandler, err := handlers.NewProcessHandler(cfg.Logger, cfg.Config)
	if err != nil {
		cfg.Logger.Errorf("Failed to initialize process handler: %v", err)
	}

	// Health check endpoint (no auth required)
	router.GET("/healthz", healthHandler.Healthz)

	// API v1 routes
	v1 := router.Group("/api/v1")
	{
		// Auth routes (no auth required)
		if authHandler != nil {
			auth := v1.Group("/auth")
			{
				auth.POST("/login", authHandler.Login)
				auth.POST("/register", authHandler.Register)
				auth.GET("/me", middleware.AuthMiddleware(jwtManager), authHandler.GetUserInfo)
			}
		}

		// Protected routes
		protected := v1.Group("/")
		protected.Use(middleware.AuthMiddleware(jwtManager))
		{
			// Metrics routes
			if metricsHandler != nil {
				metrics := protected.Group("/metrics")
				{
					metrics.GET("/nodes", metricsHandler.GetAvailableNodes)
					metrics.GET("/current", metricsHandler.GetCurrentMetrics)
					metrics.GET("/stream", metricsHandler.StreamMetrics)
					metrics.GET("/series", metricsHandler.GetMetricSeries)
				}
			}
			if queryLogHandler != nil {
				protected.GET("/query-log", queryLogHandler.ListQueryLog)
			}
			if processHandler != nil {
				processes := protected.Group("/processes")
				{
					processes.GET("", processHandler.GetCurrentProcesses)
					processes.GET("/stream", processHandler.StreamProcesses)
					processes.POST("/kill", processHandler.KillProcess)
				}
			}
		}
	}

	return router
}